pyproject.toml
[project]
name = "turbo-noodle"
version = "0.1.0"
description = "Conversational Movie AI Agent REST API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.103.0",
    "uvicorn>=0.23.2",
    "sqlalchemy>=2.0.20",
    "pydantic>=2.3.0",
    "ollama>=0.1.0",
    "pandas>=2.1.0",
    "python-dotenv>=1.0.0",
    "langgraph>=1.1.3",
    "langchain-ollama>=1.0.1",
    "langchain-community>=0.4.1",
    "pyyaml>=6.0",
    "ruff>=0.15.8",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.25.0",
    "tabulate>=0.9.0",
    "streamlit>=1.35.0",
    "requests>=2.31.0",
]

[tool.uv]
default-groups = ["dev"]

[tool.uv.pip]
generate-hashes = true
